requirements.txt
requests
dnspython
sublist3r